Disable enhanced navigation on SIdebar
Enhanced navigation attempts to preserve DOM manipulations, e.g. attrributes not added by Blazor.  In this case, the toggle for the offcanvas navigation is preserved, before it can be removed.

Thus when you land on the next page, enhanced navigation restores the toggle, which means you have to double-press the hamburger to get it to activate again.

Since navigations from the Sidebar are big navigations, I suppose it's simplest to just disable the enhanced navigation there.